LoginCheckInterceptor 부분을 Exception 처리하게 구현.
