+++ src/main/java/com/study/multiserverlogin/HomeController.java
@@ -1,8 +1,8 @@
 package com.study.multiserverlogin;
 
 import com.study.multiserverlogin.response.BaseResponse;
-import com.study.multiserverlogin.response.LoginResponse;
-import com.study.multiserverlogin.login.LoginService;
+import com.study.multiserverlogin.login.LoginServiceImpl;
+import com.study.multiserverlogin.response.login.LoginResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,20 +14,19 @@ import javax.servlet.http.HttpSession;
 @RequiredArgsConstructor
 public class HomeController {
 
-    private final LoginService loginService;
+    private final LoginServiceImpl loginService;
 
     /*
      * 현재 프로젝트에서 로그인/비로그인 구분 하는 곳
      */
     @GetMapping("/")
-    public ResponseEntity<? extends BaseResponse> homeTest(HttpSession session) {
-
-        return ResponseEntity.ok(loginService.loginCheck(session));
+    public String homeTest() {
+        return "ok";
     }
 
     @GetMapping("/login-check")
     public ResponseEntity<? extends BaseResponse> loginCheck(HttpSession session) {
-        return ResponseEntity.ok(loginService.loginCheck(session));
+        return ResponseEntity.ok(LoginResponse.success("로그인 중", null));
     }
 
 }
