+++ src/main/java/com/study/multiserverlogin/error/exHandler/ExControllerAdvice.java
@@ -1,8 +1,10 @@
 package com.study.multiserverlogin.error.exHandler;
 
+import com.study.multiserverlogin.error.exception.LoginCheckException;
 import com.study.multiserverlogin.error.exception.LoginException;
 import com.study.multiserverlogin.response.BaseResponse;
-import com.study.multiserverlogin.response.LoginResponse;
+import com.study.multiserverlogin.response.login.LoginExceptionResponse;
+import com.study.multiserverlogin.response.login.LoginResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -20,12 +22,12 @@ public class ExControllerAdvice {
      * 특정 클래스를 지정안했을때는 전체적용, default message 느낌
      * Base 적인거 몇개만 지정 해 두고 추가로 생길때마다 판단해서 추가.
      */
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ErrorResult illegalExHandler(IllegalArgumentException e) {
-        log.error("[illegalExHandler] ex", e);
-        return ErrorResult.of(ErrorStatus.BAD_REQUEST, e.getMessage());
-    }
+//    @ResponseStatus(HttpStatus.BAD_REQUEST)
+//    @ExceptionHandler(IllegalArgumentException.class)
+//    public ErrorResult illegalExHandler(IllegalArgumentException e) {
+//        log.error("[illegalExHandler] ex", e);
+//        return ErrorResult.of(ErrorStatus.BAD_REQUEST, e.getMessage());
+//    }
 
     @ResponseStatus(HttpStatus.OK)
     @ExceptionHandler(LoginException.class)
@@ -33,6 +35,11 @@ public class ExControllerAdvice {
         log.error("[loginExHandler] ex =", e);
 //        return ErrorResult.of(ErrorStatus.LOGIN_EXCEPTION, e.getMessage());
         return ResponseEntity.badRequest().body(LoginResponse.fail("로그인이 필요합니다.!!", e.getMessage()));
+    }
+
+    public ResponseEntity<? extends BaseResponse> loginCheckExHandler(LoginCheckException e){
+       log.error("[loginCheckExHandler] ex = ", e);
+        return ResponseEntity.badRequest().body(LoginExceptionResponse.loginError("로그인이 필요합니다.", e.getClass().getSimpleName()));
     }
