+++ src/main/java/com/study/multiserverlogin/interceptor/LoginCheckInterceptor.java
@@ -2,10 +2,11 @@ package com.study.multiserverlogin.interceptor;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.study.multiserverlogin.domain.session.SessionName;
-import com.study.multiserverlogin.error.exception.LoginException;
-import com.study.multiserverlogin.response.LoginResponse;
+import com.study.multiserverlogin.error.exception.LoginCheckException;
+import com.study.multiserverlogin.login.LoginService;
+import com.study.multiserverlogin.response.login.LoginResponse;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.servlet.HandlerInterceptor;
 
 import javax.servlet.http.HttpServletRequest;
@@ -13,9 +14,11 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
 @Slf4j
+@RequiredArgsConstructor
 public class LoginCheckInterceptor implements HandlerInterceptor {
 
     private final ObjectMapper objectMapper = new ObjectMapper();
+    private final LoginService loginService;
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
@@ -25,16 +28,18 @@ public class LoginCheckInterceptor implements HandlerInterceptor {
 
         HttpSession session = request.getSession();
 
-        if (session == null || session.getAttribute(SessionName.LOGIN_SESSION.name()) == null) {
+//        if (session == null || session.getAttribute(SessionName.LOGIN_SESSION.name()) == null) {
 
-            LoginResponse<String> resultMessage = LoginResponse.needLogin("로그인이 필요합니다.!!", requestURI, "");
-            String resultStringMessage = objectMapper.writeValueAsString(resultMessage);
-            response.setContentType("application/json");
-            response.setCharacterEncoding("utf-8");
-            response.getWriter().write(resultStringMessage);
+//            LoginResponse<String> resultMessage = LoginResponse.needLogin("로그인이 필요합니다.!!", requestURI, "");
+//            String resultStringMessage = objectMapper.writeValueAsString(resultMessage);
+//            response.setContentType("application/json");
+//            response.setCharacterEncoding("utf-8");
+//            response.getWriter().write(resultStringMessage);
+//
+//            return false;
 
-            return false;
-//            throw new LoginException("로그인 되지 않은 사용자.");
+        if (!loginService.isLoginCheck(session)) {
+            throw new LoginCheckException("로그인 되지 않은 사용자.");
         }
 
         return true;
