+++ src/main/java/com/study/multiserverlogin/login/LoginStatus.java
@@ -1,0 +1,5 @@
+package com.study.multiserverlogin.login;
+
+public enum LoginStatus {
+    SUCCESS, FAIL
+}
