--- src/main/java/com/study/multiserverlogin/response/LoginErrorResponse.java
@@ -1,11 +1,0 @@
-package com.study.multiserverlogin.response;
-
-public class LoginErrorResponse<T> implements BaseResponse {
-
-
-
-    String message;
-
-    T result;
-
-}
